backend/requirements/common.txt
django~=3.0.0
djangorestframework
djangorestframework-jwt
django-pydenticon
django-cors-headers
pillow

